Issue #5 mv ClientStack EndpointStack, rm Client retry - factored out to consumer repos
