+++ app.py
@@ -11,7 +11,7 @@ from stacks.control_broker_stack import (
 )
 from stacks.pipeline_stack import GitHubCDKPipelineStack
 from stacks.test_stack import TestStack
-from stacks.client_stack import ClientStack
+from stacks.endpoint_stack import EndpointStack
 
 STACK_VERSION = "V0x6x3"
 
@@ -42,9 +42,9 @@ if app.node.try_get_context("control-broker/post-deployment-testing/enabled"):
         env=env
     )
 if app.node.try_get_context("control-broker/client/enabled"):
-    ClientStack(
+    EndpointStack(
         deploy_stage or app,
-        f"ControlBrokerClientStack{STACK_VERSION}",
+        f"ControlBrokerEndpointStack{STACK_VERSION}",
         control_broker_outer_state_machine=control_broker_stack.outer_eval_engine_state_machine,
         control_broker_roles=control_broker_stack.Input_reader_roles,
         control_broker_eval_results_bucket=control_broker_stack.eval_results_reports_bucket,
